+++ pkg/namespaces/labels.go
@@ -1,8 +1,6 @@
 package namespaces
 
 const (
-	// NamespaceIDLabel allows selecting a namespace by its ID.
-	NamespaceIDLabel = `namespace.metadata.stackrox.io/id`
 	// NamespaceNameLabel allows selecting a namespace by its name.
 	NamespaceNameLabel = `namespace.metadata.stackrox.io/name`
 )