+++ CHANGELOG.md
@@ -21,6 +21,7 @@ All notable changes to this project that require documentation updates will be d
 - Added the ability to make policies based on `Image OS` (ROX-4083).
 - roxctl image scan and /v1/image/<image id> no longer return snoozed CVEs as a part of their output. The `include-snoozed` command line parameter
   and the `includeSnoozed` query parameter respectively can be used to include all CVEs.
+- The 'namespace.metadata.stackrox.io/id' label is now removed in order to better support Terraform cluster management.
 
 ## [46.0]
 - Added the following REST APIs: